+++ src/locales/en/common.json
@@ -825,7 +825,7 @@
     "undelegating": "Undelegating",
     "delegation": {
       "sectionLabel": "Delegations",
-      "addDelegation": "Add Delegation",
+      "addDelegation": "Add delegation",
       "info": {
         "title": "Earn rewards",
         "cta": "Earn rewards"
